% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Centered Linear Terms}
\usage{
center(x, center = NULL)
}
\arguments{
\item{x}{a numeric vector for which to compute the centered values. Missing values
are permitted and result in corresponding missing values in the output.}

\item{center}{an optional value to use for the center of \code{x}.}
}
\value{
The centered value of \code{x}.
}
\description{
Computes centered values. Used primarily in a linear regression formula.
}
\note{
The centering value by default is computed by the method described
in Cohn and others (1992).
}
\examples{
# trivial ceneterd values from 1 to 10
center(seq(10))
}
\references{
Cohn, T.A., Caulder, D.L., Gilroy, E.J., Zynjuk, L.D., and
Summers, R.M., 1992, The validity of a simple statistical model for
estimating fluvial constituent loads---An empirical study involving nutrient
loads entering Chesapeake Bay: Water Resources research, v. 28, no. 5, p.
937--942.
}
\seealso{
\code{\link[smwrBase]{quadratic}}, \code{\link{scale}}
}
\keyword{manip}

