% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jackStats.R
\name{jackStats}
\alias{jackStats}
\title{Jackknife Statistics}
\usage{
jackStats(fit, which = "load")
}
\arguments{
\item{fit}{an object of class "loadReg"---output from \code{loadReg}. Can also
be an object of class "censReg."}

\item{which}{a character string indicating the "load" or
"concentration" model for an object of class "loadReg" or "censReg" for
an object of class "censReg."}
}
\value{
An object of class "jackStats" containing these components:
coef, the table of coefficient estimates, the jackknife bias and standard errors\cr
coefficients, the jackknifed coefficients\cr
pctcens, the percentage of left-censored values. \cr
The PRESS statistic and individual jackknife differences are also returned
when the percentage of censoring is 0.
}
\description{
Compute selected jackknife statistics for a rating-curve load-estimation model.
}
\note{
The \code{jackStats} function can only be used when the analysis is AMLE.

Abdi and Williams (2010) describe the jackknife as refering to two related techniques: the first
estimates the parameters, their bias and standard errors and the second evaluates the
predictive performance of the model. The second technique is the PRESS statistic (Helsel
and Hirsch, 2002), but can only be used on uncensored data; it is computed by \code{jackStats}
when no data are censored. The first technique can be used to assess the coefficients of the
regression---the bias should be small and the jackknife standard errors should not be much
different from the standard errors reported for the regression. Efron and Tibshirani (1993)
suggest that the bias is small if the relative bias (biuas divided by the jackknife standard
error) is less than 0.25.
}
\examples{
# From application 1 in the vignettes
data(app1.calib)
app1.lr <- loadReg(Phosphorus ~ model(1), data = app1.calib,
 flow = "FLOW", dates = "DATES", conc.units="mg/L",
 station="Illinois River at Marseilles, Ill.")
jackStats(app1.lr)
}
\references{
Abdi, H. and Williams, L.J., 2010, Jackknife, in encyclopedia of research design,
Salkind, N.J., editor: Thousand Oaks, Calif., SAGE Publications, 1719 p.

Efron, B. and Tibshirani, R.J., 1993, An introduction to the bootstrap: Boca Raton,
Fla., Chapman and Hall/CRC, 436 p.

Helsel, D.R. and Hirsch, R.M., 2002, Statistical methods in water resources:
U.S. Geological Survey Techniques of Water-Resources Investigations, book 4,
chap. A3, 522 p.
Salkind,
}
\seealso{
\code{\link{loadReg}}
}
\keyword{utilities}

