% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadReport.R
\name{loadReport}
\alias{loadReport}
\title{Create Load Report}
\usage{
loadReport(x, file)
}
\arguments{
\item{x}{the load model.}

\item{file}{the output file base name; the .pdf suffix 
is appended to make the actual file name. if missing, then the
name of \code{x} is used as the base name.}
}
\value{
The actual file name is returned invisibly.
}
\description{
Create a 2-page pdf file report of a rating-curve load model. The 
report contains the text output and 6 diagnostic plots.
}
