% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makepredictcall.center.R
\name{makepredictcall.center}
\alias{makepredictcall.center}
\title{Utility Function for Safe Prediction}
\usage{
\method{makepredictcall}{center}(var, call)
}
\arguments{
\item{var}{a variable.}

\item{call}{the term in the formula, as a call.}
}
\value{
A replacement for \code{call} for the prediction variable.
}
\description{
A utility to help \code{\link{model.frame.default}} create the right matrices 
when predicting from models with \code{center} term. Used only internally.
}
