% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Models.R
\docType{data}
\name{Models}
\alias{Models}
\title{Models
A dataset that describes the equivalent formula for each of the
predefined model number.}
\format{Data frame with 9 rows and 2 columns\cr
\tabular{lll}{
Name \tab Type \tab Description\cr
Number \tab integer \tab The predefined model number\cr
Formula \tab factor \tab The equivalent formula for the predefined model number\cr
}}
\usage{
Models
}
\description{
Models
A dataset that describes the equivalent formula for each of the
predefined model number.
}
\examples{
data(Models)
print(Models)
}
\seealso{
\code{\link{model}}
}
\keyword{datasets}

