% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predLoad.R
\name{predLoad}
\alias{predLoad}
\title{Predict Loads}
\usage{
predLoad(fit, newdata, load.units = fit$load.units, by = "total",
  seopt = "exact", allow.incomplete = FALSE, conf.int = 0.95,
  print = FALSE)
}
\arguments{
\item{fit}{the output from \code{loadReg}.}

\item{newdata}{a data frame of the prediction variables. Missing values
are not permitted in any column in \code{newdata}. Observations with
missing values \code{NAs} must be removed before prediction. Columns that
are not needed for prediction that contain missing values can be removed
before removing all rows with missing values. The maximum number of rows
permitted in \code{newdata} is 176000.}

\item{load.units}{a character string indicating the units of the
predicted loads/fluxes. By default, uses the value specified in
\code{loadReg}. See \code{\link{loadReg}} for a complete list of options.}

\item{by}{the time frame for estimates. See \code{Details}.}

\item{seopt}{a character string indicating how to comute the
standard error of the aggregated load estimates, must be either
"exact" or "approximate." Only the first letter is necessary.}

\item{allow.incomplete}{compute loads for periods withing
missing values or incomplete record? See \code{Details}.}

\item{conf.int}{the confidence interval to compute for loads
computed by "day" or "unit." The confidence interval is fixed at
0.95 for any other value for \code{by}. See \code{Details}.}

\item{print}{print a report summary of the load estimate?}
}
\value{
A data frame containing the load estimates.
}
\description{
Estimate loads from a rating-curve model from \code{loadReg}
for a new data frame, aggregating the loads by specified time
periods.
}
\details{
The time frame specified by \code{by} can be "unit,"
"day," "month," "water year," "calendar year," "total," or
the name of a column in \code{newdata} that can be used to
group the data.

If \code{allow.incomplete} is \code{TRUE}, then loads will be
computed based on all nonmissing values, otherwise missing values
\code{NAs} will be returned. For this application, missing values
includes \code{NAs} and gaps in the record, except for \code{by}
set to "total" or user defined groups where missing values only
includes \code{NAs}. For prediction by "day" when there are variable
number of unit values per day, \code{allow.incomplete} must be
set to \code{TRUE}.

The term confidence interval is used here as in the original
documentation for LOADEST, but the values that are reported are
the prediction intervals, computed from the SEP.
}
\examples{
# From application 1 in the vignettes
data(app1.calib)
app1.lr <- loadReg(Phosphorus ~ model(1), data = app1.calib,
 flow = "FLOW", dates = "DATES", conc.units="mg/L",
 station="Illinois River at Marseilles, Ill.")
predLoad(app1.lr, app1.calib)
}
\seealso{
\code{\link{loadReg}},
}

