% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.jackStats.R
\name{print.jackStats}
\alias{print.jackStats}
\title{Print Results}
\usage{
\method{print}{jackStats}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class "jackStats"---output from \code{jackStats}.}

\item{digits}{the number of significant digits to print.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
The object \code{x} is returned invisibly.
}
\description{
Print the results of a jackknife analysis of a censored regression.
}
\note{
The printed output includes the original original estimate,
the jackknife bias and standard error and the relative bias for
each parameter in the regression model.
}
\seealso{
\code{\link{loadReg}}
}
\keyword{utilities}
