% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resampleUVdata.R
\name{resampleUVdata}
\alias{resampleUVdata}
\title{Resample Unit-Value Data}
\usage{
resampleUVdata(UVdata, time.step = 15, start.date = "", end.date = "",
  max.diff = "2 hours", missing.days = "exclude")
}
\arguments{
\item{UVdata}{the dataset containing the unit-values data. Must have one column that
represents the time of the observation that is class "POSIXt." Missing values are not
permitted in that column.}

\item{time.step}{the time step of the new data in minutes; must divide an hour
exactly evenly. The default value is 15 minutes.}

\item{start.date}{a character string indicating the first day of the output dataset. The
default value ("") indicates use the first day in \code{UVdata}.}

\item{end.date}{a character string indicating the last day of the output dataset. The
default value ("") indicates use the last day in \code{UVdata}.}

\item{max.diff}{a character string indicating the maximum difference in time to
sucessfully resample the unit-value data. The default is "2 hours" see
\code{\link[smwrBase]{mergeNearest}} for details.}

\item{missing.days}{a characer string indicating what action should be taken for days
not present in \code{UVdata}. Must be either "exclude" to remove those days from the output,
or "include" to retain them. Can be abbreviated. If \code{missing.days} is "include,"
then partial days within \code{max.diff} will be included in the output data frame.}
}
\value{
A data frame like \code{UVdata} but having a uniform time step.
}
\description{
Unit-value data can be recorded at any arbitrary intervals. For some
applications, such as load estimates, a uniform series is required. The
\code{resampleUVdata} function resamples the orginal unit-value data to a consistent
time interval.
}

