% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seg.R
\name{seg}
\alias{seg}
\title{Load Model}
\usage{
seg(x, N)
}
\arguments{
\item{x}{the data to segment. Missing values are permitted and result
corresponing in missing values in output.}

\item{N}{the number of breaks in the segmented model.}
}
\value{
The data in \code{x} with attributes to build the segmented model.
}
\description{
Support function for building a segmented rating curve load model. Required
in the formula in \code{segLoadReg} to define the segmented model.
}

