% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/segLoadReg.R
\name{segLoadReg}
\alias{segLoadReg}
\title{Load Estimation}
\usage{
segLoadReg(formula, data, subset, na.action, flow, dates, flow.units = "cfs",
  conc.units = "", load.units = "kg", time.step = "day", station = "",
  trace = TRUE)
}
\arguments{
\item{formula}{a formula describing the regression model. See \bold{Details}.}

\item{data}{the data to search for the variables in \code{formula}.}

\item{subset}{an expression to select a subset of the data.}

\item{na.action}{what to do with missing values.}

\item{flow}{character string indicating the name of the
flow column.}

\item{dates}{character string indicating the name of the
date column.}

\item{flow.units}{character string describing the flow unit.}

\item{conc.units}{character string describing the concentration
unit.}

\item{load.units}{character string describing the load unit.}

\item{time.step}{character string describing the time step of
the calibration data.}

\item{station}{character string description of the station.}

\item{trace}{if logical, then if \code{TRUE} print a short summary of the
segmented fit. Otherwise a character string and the segmented model is saved
as that object name.}
}
\value{
An object of class "loadReg."
}
\description{
Build a segmented rating-curve (regression) model for river load estimation.
}
\details{
The left-hand side of the formula can be any numeric variable, just
as with \code{lm} or a variable of class "lcens," "mcens," "qw," or "Surv."
The response variable must be a column in \code{data}; it cannot be constructed using
\code{as.lcens}, \code{as.mcens}, or \code{Surv}. The initial segmented model is
based on uncensored data---simple substitution of 1/2 the reporting limit is used
for left-censored values and multiply censored values cause the analysis to fail.

The first term of right-hand side must be defined by the \code{seg} function
with the number of segments. The first term may be followed by any number of
additional terms. The final model will place the segmeted term in the last postition
and \code{seg} will be replaced by the proper call to \code{segment}.
}
\examples{
# From application 1 in the vignettes
data(app1.calib)
app1.lr <- loadReg(Phosphorus ~ model(1), data = app1.calib,
 flow = "FLOW", dates = "DATES", conc.units="mg/L",
 station="Illinois River at Marseilles, Ill.")
print(app1.lr)
}
\references{
will need some.
}
\seealso{
\code{\link{censReg}}, \code{link{seg}}, \code{\link{segment}}
}
\keyword{censored}
\keyword{loads}
\keyword{regression}

