% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Segmented Model}
\usage{
segment(x, psi)
}
\arguments{
\item{x}{the data to segment. Missing values are permitted and result
corresponding in missing values in output.}

\item{psi}{a numeric vector of the breakpoints.}
}
\value{
A matrix contining a column named X of the data in \code{x} and
paired U and P columns for each of the breaks in \code{psi} that form the
basis for that segment.
}
\description{
Computes the basis for a segmented model. Used primarily in a linear
regression or load model.
}

