% Generated by roxygen2 (4.0.1): do not edit by hand
\name{selBestSubset}
\alias{selBestSubset}
\title{Load Estimation}
\usage{
selBestSubset(formula, min.formula = ~1, data, subset, na.action, flow, dates,
  flow.units = "cfs", conc.units = "", load.units = "kg",
  time.step = "day", station = "", criterion = c("AIC", "SPPC"))
}
\arguments{
\item{formula}{a formula describing the regression model. See \code{\link{loadReg}}
for details.}

\item{min.formula}{a formula containing the minimum variables to use in the final
model. The default is to only force the intercept term, whihc will normally be
acceptable. In some rare cases, the log of flow may be needed.}

\item{data}{the data to search for the variables in \code{formula}.}

\item{subset}{an expression to select a subset of the data.}

\item{na.action}{what to do with missing values.}

\item{flow}{character string indicating the name of the
flow column.}

\item{dates}{character string indicating the name of the
date column.}

\item{flow.units}{character string describing the flow unit.}

\item{conc.units}{character string describing the concentration
unit.}

\item{load.units}{character string describing the load unit.}

\item{time.step}{character string describing the time step of
the calibration data.}

\item{station}{character string description of the station.}

\item{criterion}{the criterion to use for subset selection, must be either "AIC"
or "SPPC."}
}
\value{
An object of class "loadReg."
}
\description{
Select the "best" subset of a user-defined rating-curve (regression)
model for rver load estimation.
}
\note{
The printed output of the model inlcudes the \code{anova} component from
\code{step}. That table summarizes the step wise selection and the criterion used
for each step. The statistics possibly represent a smaller  sample size than used
for the final model because the \code{step} function requires a data set with no
missing values. If missing values are found a warning is printed.
}
\seealso{
\code{\link{censReg}}, \code{\link[stats]{step}}
}
\keyword{censored}
\keyword{loads}
\keyword{regression}

