% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vif.loadReg.R
\name{vif.loadReg}
\alias{vif.loadReg}
\title{Variance Inflation Factors}
\usage{
\method{vif}{loadReg}(model, ...)
}
\arguments{
\item{model}{an object of class "loadReg"---output from \code{loadReg}.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
further arguments passed to or from other methods.
}
\description{
Computes the variance inflation factor (Helsel and Hirsch, 2002) for each
variable in a load regression.
}
\examples{
# From application 1 in the vignettes
data(app1.calib)
app1.lr <- loadReg(Phosphorus ~ model(1), data = app1.calib,
 flow = "FLOW", dates = "DATES", conc.units="mg/L",
 station="Illinois River at Marseilles, Ill.")
vif(app1.lr, app1.calib)
}
\seealso{
\code{\link{loadReg}}
}
\keyword{regression}

