% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadConvFactor.R
\name{loadConvFactor}
\alias{loadConvFactor}
\title{Unit Conversion}
\usage{
loadConvFactor(flow.units, conc.units, load.units)
}
\arguments{
\item{flow.units}{character string describing the flow unit. The only valid values are
"cubic meter per second," "cms," "cubic feet per second," or "cfs."}

\item{conc.units}{character string describing the concentration unit. The valid units
are "mg/l," "mg/L," "ug/l," "ug/L," "ng/l," "ng/L," "milligrams per liter,"
"micrograms per liter," "nanograms per liter," "col/100mL," "col/dL," or
"colonies per 100mL."}

\item{load.units}{character string describing the load unit. The valid values are
"pounds," "tons," "mg," "milligrams," "grams," "g," "kilograms," "kg,"
"metric tons," "Mg," or "million colonies."}
}
\value{
The conversion factor.
}
\description{
Computes the conversion factor to compute the flux units
(load units) from the concentration and flow units
}
\examples{
loadConvFactor("cubic meter per second","milligrams per liter","pounds")
}
\keyword{conversions}
\keyword{unit}

