% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadReg.R
\name{loadReg}
\alias{loadReg}
\title{Load Estimation}
\usage{
loadReg(formula, data, subset, na.action, flow, dates, flow.units = "cfs",
  conc.units = "", load.units = "kg", time.step = "day", station = "")
}
\arguments{
\item{formula}{a formula describing the regression model. See \bold{Details}.}

\item{data}{the data to search for the variables in \code{formula}.}

\item{subset}{an expression to select a subset of the data.}

\item{na.action}{what to do with missing values.}

\item{flow}{character string indicating the name of the
flow column.}

\item{dates}{character string indicating the name of the
date column.}

\item{flow.units}{character string describing the flow units. See \bold{Details}.}

\item{conc.units}{character string describing the concentration
unit. See \bold{Details}.}

\item{load.units}{character string describing the load unit. See \bold{Details}.}

\item{time.step}{character string describing the time step of
the calibration data. Must be one of "instantaneous," "2 hours," "3 hours,"
"4 hours," "6 hours," "12 hours," or "day." The default is "day."}

\item{station}{character string description of the station.}
}
\value{
An object of class "loadReg."
}
\description{
Build a rating-curve (regression) model for river load estimation.
}
\details{
The left-hand side of the formula may be any numeric variable, just as with
\code{lm} or a variable of class "lcens," "mcens," or "qw." Also permitted are
variables constructed using \code{Surv} of \code{type} "right,", "interval," or
"interval2" (for left-censored data, use \code{as.lcens}.\cr
For un- or left-censored data, AMLE is used unless weights are specified in
the model, then MLE is used, through a call to \code{survreg}. For any other
censored data, MLE is used.\cr

Typically, \code{loadReg} expects the response variable to have units of
concentration, mass per volume. For these models, See \code{\link{loadConvFactor}}
for details about valid values for \code{flow.units}, \code{conc.units} and
\code{load.units}. For some applications, like bed load estimation, the response
variable can have units of flux, mass per time. For these models, \code{conc.units}
can be expressed as any valid \code{load.units} per day. The rate must be expressed
in terms of "/d," "/dy," or "/day."
}
\examples{
# From application 1 in the vignettes
data(app1.calib)
app1.lr <- loadReg(Phosphorus ~ model(1), data = app1.calib,
 flow = "FLOW", dates = "DATES", conc.units="mg/L",
 station="Illinois River at Marseilles, Ill.")
print(app1.lr)
}
\references{
Runkel, R.L., Crawford, C.G., and Cohn, T.A., 2004, Load estimator (LOADEST):
a FORTRAN program for estimating constituent loads in streams and rivers: U.S.
Geological Survey Techniques and Methods book 4, chap. A5, 69 p.
}
\seealso{
\code{\link{censReg}}, \code{\link{as.lcens}}, \code{\link{as.mcens}},
\code{\link{Surv}}
}
\keyword{censored}
\keyword{loads}
\keyword{regression}

