% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rmse.loadReg.R
\name{rmse.loadReg}
\alias{rmse.loadReg}
\title{Root-Mean-Squared Error}
\usage{
\method{rmse}{loadReg}(x, model = c("load", "concentration"), ...)
}
\arguments{
\item{x}{the output from \code{loadReg}.}

\item{model}{the type of model, must be either "load" or "concentration."}

\item{\dots}{not used, required for other methods.}
}
\value{
The estimated root-mean-squared error, also know as the residual
standard error.
}
\description{
Compute the root-mean-squared error (RMSE) of the difference between
observed values and the fitted values for the load or concentration model.
The RMSEs will be the same unless the log of flow is not an
explanatory variable.
}
\examples{
# From application 1 in the vignettes
data(app1.calib)
app1.lr <- loadReg(Phosphorus ~ model(1), data = app1.calib,
 flow = "FLOW", dates = "DATES", conc.units="mg/L",
 station="Illinois River at Marseilles, Ill.")
rmse(app1.lr)
}
\seealso{
\code{\link{loadReg}},
}

